feat: 이상증상 리스트 get, add api에 accessToken, dogId 가져오는 이벤트 추가

+++ app/src/main/java/com/project/meongcare/symptom/model/entities/SymptomData.kt
@@ -8,7 +8,7 @@ data class ResultSymptom(
 )
 
 data class ToAddSymptom(
-    val dogId: Int,
+    val dogId: Long,
     val symptomString: String,
     val note: String,
     val dateTime: String,
